+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Enhancements
 - **General Changes**
   - Expanded the race breakdown on the demographics page
+  - Added a data quality report page
 - **Gem Changes**
   - Dropped support for Ruby 2.2
   - Updated to Ruby 2.3.3